# always inline vm_setivar