It seems like GCC version 9.3.0 isn't inlining this function even though it says `static inline`.  Since the compiler isn't inlining `vm_setivar`, the `is_attr` variable test doesn't get removed, and we can see that in a report from perf:

<img width="1033" alt="ruby — perf :home:aaron:git:ruby — ssh whiteclaw local — 128×47 2020-11-19 14-25-35" src="https://user-images.githubusercontent.com/3124/99731816-47414e80-2a73-11eb-9420-0e60f633f55b.png">

`is_attr` is a hardcoded number [here](https://github.com/ruby/ruby/blob/44ad72fa2113d0e111a1cf76327bcc13de8393e1/vm_insnhelper.c#L1273) and [here](https://github.com/ruby/ruby/blob/44ad72fa2113d0e111a1cf76327bcc13de8393e1/vm_insnhelper.c#L2696), so this should be trivial for the compiler to remove.

Forcing this function to always be inline shows a small speed improvement on optcarrot:

![iv-set](https://user-images.githubusercontent.com/3124/99732120-c6cf1d80-2a73-11eb-809d-b3ee3f288e24.png)