@k0kubun 2020-11-19
😄 https://github.com/ruby/ruby/pull/1590