@k0kubun 2020-11-20
JFYI, in my environment using GCC 9.3.0 (Intel 4.0GHz i7-4790K 8 cores, 16GB RAM, x86-64 Ubuntu 20.04), it became slower for some reason:

```
$ benchmark-driver benchmark.yml -v --rbenv 'before::ruby;after::after' --repeat-count=12 --output=all
before: ruby 3.0.0dev (2020-11-19T06:41:53Z master 44ad72fa21) [x86_64-linux]
after: ruby 3.0.0dev (2020-11-19T21:52:59Z always-inline-ivset 93fe04ed3b) [x86_64-linux]
Calculating -------------------------------------
                                       before                 after
Optcarrot Lan_Master.nes    47.47188177204166     46.74959032857574 fps
                            47.73454400117186     46.79469591375146
                            48.07891994463598     46.81045448101272
                            48.10177257531159     47.41038922840676
                            48.15326246264417     47.42468070415044
                            48.22981907621332     47.42525123074406
                            48.28130646116692     47.60157960284047
                            48.41794108552664     47.60475284212768
                            48.46746886619965     47.80315933736141
                            48.53849826538023     47.81169942193240
                            48.56456893926927     47.89032528008438
                            49.59060967990376     48.06008022389269
```