@k0kubun 2020-11-22
What CPU did you use for benchmarking it? I'm interested in the difference between your environment and mine. I'm thinking it might come from something like L1 cache size.