@tenderlove 2020-12-02
> What CPU did you use for benchmarking it? I'm interested in the difference between your environment and mine. I'm thinking it might come from something like L1 cache size.

I'm using a `AMD Ryzen 7 3700X 8-Core Processor` (32G ram, Ubuntu 20.04).  I think it has a bigger L1 cache size than then Intel i7.  According to [this](https://www.techpowerup.com/cpu-specs/ryzen-7-3700x.c2130), the Ryzen has 64k L1 cache, and according to [this](https://www.guru3d.com/articles-pages/core-i7-4790k-processor-review,4.html), the i7-4790K has a 32k cache.  So I think we can blame larger code size on the performance reduction on the i7-4790K.